12: moves core components under Core namespace
