+++ ImitationLib/Elements/Core/Accumulator.cs
