+++ ImitationLib/ImitationLib.csproj
@@ -46,20 +46,20 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Elements\Accumulators\Accumulator.cs" />
     <Compile Include="Elements\Accumulators\Queue.cs" />
-    <Compile Include="Elements\Collectors\Collector.cs" />
     <Compile Include="Elements\Collectors\Exit.cs" />
-    <Compile Include="Elements\Executors\Executor.cs" />
+    <Compile Include="Elements\Core\Accumulator.cs"/>
+    <Compile Include="Elements\Core\Collector.cs"/>
+    <Compile Include="Elements\Core\Element.cs"/>
+    <Compile Include="Elements\Core\Executor.cs"/>
+    <Compile Include="Elements\Core\Generator.cs"/>
+    <Compile Include="Elements\Core\IEnterable.cs"/>
+    <Compile Include="Elements\Core\IExitable.cs"/>
     <Compile Include="Elements\Executors\Service.cs" />
     <Compile Include="Elements\Generators\Enter.cs" />
-    <Compile Include="Elements\Generators\Generator.cs" />
-    <Compile Include="Elements\IEnterable.cs" />
-    <Compile Include="Elements\IExitable.cs" />
     <Compile Include="Model.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Elements\Element.cs" />
     <Compile Include="Utils\Constants.cs" />
     <Compile Include="Utils\Logger.cs" />
     <Compile Include="Utils\Transact.cs" />
@@ -68,7 +68,6 @@
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
