+++ ImitationLib/Model.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using ImitationLib.Elements;
+using ImitationLib.Elements.Core;
 using ImitationLib.Utils;
